pyproject.toml
[project]
name = "audio-transcription-app"
version = "0.1.0"
description = "Audio transcription app with speaker recognition and Streamlit GUI"
authors = [
    {name = "User",email = "user@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "streamlit (>=1.42.2,<2.0.0)",
    "pydub (>=0.25.1,<0.26.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "speechrecognition[faster-whisper] (>=3.14.1,<4.0.0)",
    "faster-whisper (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
app = "run_app:main"

